created request to random api, save words to array This is the correct one
